Move roundtrip function to tests. Output based on json diff
